pyproject.toml
[project]
name = "ai-weekly-report"
version = "0.1.0"
description = "AI Weekly Trend Report Generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.52.0",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.0",
    "jinja2>=3.1.0",
    "requests>=2.31.0",
]